Stabilize the main_loop
The assumption of FTQ is that the work that we're counting takes a
constant amount of cpu time.  To that end, we try avoiding cache misses,
compiler tricks, etc.: stuff like the "volatile" on count, the trickery
with the loop iterations.  Even the function alignment to make sure we
don't have icache / decoder / front-end issues when compiling for
different OSes.

However, with all that, we've created a weird integer-register workload
that the *hardware*, particularly intel's newer chips, can attempt to
speed up at runtime.  This speedup can cause each iteration to vary
dramatically in cycles, to the tune of 3-4x differences for *several*
intervals.  i.e. for O(msec).

On a quiet machine, it's even worse: I've seen drops in IPC that last
for O(sec).

Anyway, manually unrolled floats seems to be the stablest I've found so
far.

I'll do a more in-depth writeup in another commit.  (Probably).

Tested: FTQ on an isolated CPU (nosmt, nohz, isolcpus, etc) on Milan and
Sapphire Rapids.

Signed-off-by: Barret Rhoden <brho@google.com>